MPI enabled in extract_averages_retro.py

+++ scripts/wrf_hydro/extract_average_retro.py
@@ -21,6 +21,13 @@ from calendar import monthrange
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))+'/utils')
 from utilities import config, find_last_time
 
+from mpi4py import MPI
+
+# MPI setup
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+size = comm.Get_size()
+
 ## main function
 def main(argv):
 
@@ -41,32 +48,34 @@ def main(argv):
     site_names.remove('TRF1'); site_names.remove('TRF2')
     site_names.append('TRF')
 
-    # daily output
-    fnins = ''
-    tmp_out = f'tmp_out_{t1:%Y%m}-{t2:%Y%m}'
-    t = t1
-    while t<=t2:
-        fnins += f' 1km_daily/{t:%Y/%Y%m}.LDASOUT_DOMAIN1'
-        t += relativedelta(months=1)
-    cmd = f'cdo --sortname -f nc4 -z zip mergetime {fnins} {tmp_out}'
-    if not os.path.isfile(tmp_out):
+    tmp_out = f'tmp_out_avg_{t1:%Y%m}-{t2:%Y%m}'
+    tmp_for = f'tmp_for_avg_{t1:%Y%m}-{t2:%Y%m}'
+
+    if rank==0:
+        dout = f'basins/{t1:%Y%m}-{t2:%Y%m}/averaged'
+        if not os.path.isdir(dout):
+            os.system(f'mkdir -p {dout}')
+        # daily output
+        fnins = ''
+        t = t1
+        while t<=t2:
+            fnins += f' 1km_daily/{t:%Y/%Y%m}.LDASOUT_DOMAIN1'
+            t += relativedelta(months=1)
+        cmd = f'cdo --sortname -f nc4 -z zip mergetime {fnins} {tmp_out}'
         print(cmd); os.system(cmd)
     
-    # daily forcing
-    fnins = ''
-    tmp_for = f'tmp_for_{t1:%Y%m}-{t2:%Y%m}'
-    t = t1
-    while t<=t2:
-        fnins += f' ../forcing/1km_daily/{t:%Y%m}.LDASIN_DOMAIN1.daily'
-        t += relativedelta(months=1)
-    cmd = f'cdo --sortname -f nc4 -z zip mergetime {fnins} {tmp_for}'
-    print(cmd); os.system(cmd)
+    if rank==size-1:
+        # daily forcing
+        fnins = ''
+        t = t1
+        while t<=t2:
+            fnins += f' ../forcing/1km_daily/{t:%Y%m}.LDASIN_DOMAIN1.daily'
+            t += relativedelta(months=1)
+        cmd = f'cdo --sortname -f nc4 -z zip mergetime {fnins} {tmp_for}'
+        print(cmd); os.system(cmd)
 
-    dout = f'basins/{t1:%Y%m}-{t2:%Y%m}/averaged'
-    if not os.path.isdir(dout):
-        os.system(f'mkdir -p {dout}')
-    
-    for name in site_names:
+    comm.Barrier()    
+    for name in site_names[rank::size]:
         
         fnout = f'{dout}/{name}_SWE.txt'
         cmd = f'{cdocmd} -mul -selname,SNEQV {tmp_out} ../../domain/masks/{name}.nc > {fnout}'
@@ -94,7 +103,9 @@ def main(argv):
         df_swe['PREC']  = df_prec['PREC']
         df_swe.to_csv(fnout, index=True, float_format='%.3f', date_format='%Y-%m-%d')
     
-    os.system(f'rm -f {tmp_out} {tmp_for} {dout}/*.txt')
+    comm.Barrier()
+    if rank==0:
+        os.system(f'rm -f {tmp_out} {tmp_for} {dout}/*.txt')
     
     return 0
